# Make GeoRaster2.resize faster