Before (in-memory raster):

```python
In [1]: import telluric as tl

In [2]: r = tl.GeoRaster2.open("creaf_gmap.tif", lazy_load=False)

In [3]: %time r.resize(ratio=0.5)
CPU times: user 3.38 s, sys: 23 ms, total: 3.4 s
Wall time: 3.4 s
```
Before (on-disk raster):

```python
In [1]: import telluric as tl

In [2]: r = tl.GeoRaster2.open("creaf_gmap.tif", lazy_load=True)

In [3]: %time r.resize(ratio=0.5)
CPU times: user 193 ms, sys: 346 µs, total: 193 ms
Wall time: 193 ms
```

After (in-memory raster):

```python
In [1]: import telluric as tl

In [2]: r = tl.GeoRaster2.open("creaf_gmap.tif", lazy_load=False)

In [3]: %time r.resize(ratio=0.5)  # 50x faster
CPU times: user 58.1 ms, sys: 8.62 ms, total: 66.7 ms
Wall time: 65.8 ms
```

After (on-disk raster):

```python
In [1]: import telluric as tl

In [2]: r = tl.GeoRaster2.open("creaf_gmap.tif", lazy_load=True)

In [3]: %time r.resize(ratio=0.5)  # 2.5x faster
CPU times: user 64.9 ms, sys: 11.1 ms, total: 75.9 ms
Wall time: 75.2 ms
```